+++ contracts/assets/AssetAdapter.sol
@@ -6,7 +6,7 @@ import "../dependencies/ERC20Protocol.sol";
 /// @author Melonport AG <team@melonport.com>
 /// @notice This is to be considered as a protocol on how to access the underlying Asset Contract
 /// @notice This extends the ERC20 Protocol
-contract AssetInterface is ERC20Protocol {
+contract AssetAdapter is ERC20Protocol {
     // CONSTANT METHODS
 
     function getName() constant returns (string) {}