+++ contracts/exchange/ExchangeAdapter.sol
@@ -5,7 +5,7 @@ import '../dependencies/ERC20.sol';
 /// @title Exchange Protocol Contract
 /// @author Melonport AG <team@melonport.com>
 /// @notice This is to be considered as a protocol on how to access the underlying Exchange Contract
-contract ExchangeAdaptor {
+contract ExchangeAdapter {
 
     // CONSTANT METHODS